First attempt at adding mirrors - messed up
This one is the first attempt at adding mirrors - you can see that the
result is not a viable puzzle as indicated by the orange beams. Also
somehow the D, T and O characters are not hit but not signalled as so -
so something is wrong with marking cells visited too.

The notable error here is that though we move the lasers after inserting
the mirror, we don't change the *orientation* of the lasers!

There also seems to be an issue with interpretation of mirror direction.