Final version, with timings and randomness tweaked.
This is the version I'm pushing for testing (after play-testing).
Modifications:
- Bumped up POSITION_PERTURBATION_AMPLITUDE from 40 to 60
- Added xyPerturbationGain - Cheese sets this to 0.5 (default is 1.0) so
that it doesn't wiggle around so much.
- Widened the hole (because mice wiggle around more so weren't really
entering the hole sometimes)
- Attempted to add logic of releasing mice from the hole based on a
combination of the previous min-mice-in-field as well as a new
dormantDelay that is supposed to be dependent on the difference between
this rat and the rat with the longest path left. However I haven't
calculated the dormantDelay properly. It seems to "work" in that the
(longest path) brown mouse no longer ends up alone at the end, so I say
"ship it".
Stop saving frames when the rats are all done moving - otherwise it
keeps endlessly generating frames if you leave it to run!