Create info table for each puzzle type
Info table is a CSV file that contains Instance ID and solution for all
puzzles of a particular type. It is used by PuzzleGeneration to pick
instance IDs when generating a puzzle packet. Also, the solution field
is a kind of end-to-end record keeping of which puzzle goes to which
solution.