Major overhaul - reformatted, and now based on names, not species.
- Now the  unique names as the primary key. These names are
alphebetical, but they are selected at random from a very large set
(some 3000 names), while checking that they are not too similar to each
other.
- animals.txt: Added optional max age (following comma) . These values
are used when generating random ages.
- major formatting changes - larger and centered numbers; slightly
narrower so that it fits nicely on the rounded-edge cards;  Cues to help
with cutting evenly into equal sized tiles.
- encoded text is now saved to a file: "qr_animals_key.txt"
- Added utils.pde for generating permuations of indices.