Ability to print out selected animal cards
I messed up while cutting up the individual cards and lost 12 of them
that made the sequence [2, 10, 14, ..., 42, 46]. I tried fixing the
printout ones but messed up even further. So I added code to re-print
selected cards. Call genAnimalTiles with null to generate all tiles,
else call it with an array of IDs to print. Note that IDs are 1-based.